+++ linux/scripts/01-core/build-helpers.sh
@@ -117,6 +117,7 @@ append_runtime_base_parent_build_arg() {
 append_runtime_accelerator_build_args() {
   append_optional_build_arg "$1" ENABLE_NVIDIA "${ENABLE_NVIDIA:-}"
   append_optional_build_arg "$1" ENABLE_AMD "${ENABLE_AMD:-}"
+  append_optional_build_arg "$1" ENABLE_ROCM_ASAN "${ENABLE_ROCM_ASAN:-}"
 }
 
 append_common_build_args() {